refactor(insights): share trends quill legend config + restore align/gap
Extract the quill legend wiring (hidden-keys, toggle persistence, isolate
context menu) into a reusable useTrendsLegendConfig hook so the trends-family
charts can share it. Restore align/gap on ChartLegendConfig and snapshot the
built-in legend at each position around the chart.

Generated-By: PostHog Code
Task-Id: d0ad71b3-9daf-4160-b003-8e2859dd0049
