+++ packages/quill/packages/charts/src/components/Legend/Legend.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react'
-import { useMemo, useState } from 'react'
+import { useMemo, useState, type ReactNode } from 'react'
 
 import { DAYS, SERIES } from '../../charts/time-series-fixtures'
 import { TimeSeriesBarChart } from '../../charts/TimeSeriesBarChart/TimeSeriesBarChart'
@@ -69,21 +69,6 @@ export const ManyItemsWraps: Story = {
     ),
 }
 
-// `renderItem` wraps each row so consumers can augment it (here a native-title tooltip on
-// right-click) while keeping the default swatch/label rendering.
-export const RenderItemWrapper: Story = {
-    render: () => (
-        <div className="w-[480px]">
-            <Legend
-                items={LIFECYCLE}
-                renderItem={(node, item) => (
-                    <span title={`Right-click ${item.label} for options`}>{node}</span>
-                )}
-            />
-        </div>
-    ),
-}
-
 function ChartLegendStory({
     show = true,
     position,
@@ -123,9 +108,16 @@ export const LegendHidden: Story = {
     render: () => <ChartLegendStory show={false} position="top" />,
 }
 
-// The chart owns the toggle state — `config.legend.show` is all that's needed; clicking a row
-// hides that series (dimmed in the legend) and the axes rescale into the freed space.
-function BuiltInToggleStory(): JSX.Element {
+// The chart owns the toggle state — `config.legend` is all that's needed; clicking a row hides
+// that series (dimmed in the legend) and the axes rescale into the freed space. These stories
+// snapshot the built-in legend at each position around the chart, not the legend in isolation.
+function BuiltInToggleStory({
+    position,
+    renderItem,
+}: {
+    position: 'top' | 'bottom' | 'left' | 'right'
+    renderItem?: (node: ReactNode, item: LegendItem) => ReactNode
+}): JSX.Element {
     const theme = useReactiveTheme()
     return (
         <Stage width={520} height={320}>
@@ -135,11 +127,25 @@ function BuiltInToggleStory(): JSX.Element {
                 theme={theme}
                 config={{
                     yAxis: { showGrid: true },
-                    legend: { show: true, position: 'bottom' },
+                    legend: { show: true, position, renderItem },
                 }}
             />
         </Stage>
     )
 }
 
-export const BuiltInToggle: Story = { render: () => <BuiltInToggleStory /> }
+export const BuiltInToggleTop: Story = { render: () => <BuiltInToggleStory position="top" /> }
+export const BuiltInToggleBottom: Story = { render: () => <BuiltInToggleStory position="bottom" /> }
+export const BuiltInToggleLeft: Story = { render: () => <BuiltInToggleStory position="left" /> }
+export const BuiltInToggleRight: Story = { render: () => <BuiltInToggleStory position="right" /> }
+
+// `renderItem` wraps each row so a consumer can augment it (here a native-title tooltip standing in
+// for a right-click context menu) while keeping the default swatch/label/toggle rendering.
+export const BuiltInToggleRenderItem: Story = {
+    render: () => (
+        <BuiltInToggleStory
+            position="bottom"
+            renderItem={(node, item) => <span title={`Right-click ${item.label} for options`}>{node}</span>}
+        />
+    ),
+}
